+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## 1.2.9
 - Anvil grid scales with scene length unit
 - Fix cube cut and box builder snap/overlay not aligning with Blender grid for imperial units
+- Don't show cube cut and box builder grid overlay in ortho views (grid is already there)
 - Fix initial texture application setting incorrect UV scale
 - Improve reliability of image texture preview